tests/wpt/tests/battery-status/battery-disabled-by-permissions-policy.https.sub.html

<!DOCTYPE html>
<body>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=/permissions-policy/resources/permissions-policy.js></script>
<script>
'use strict';

const same_origin_src = '/permissions-policy/resources/permissions-policy-battery.html';
const cross_origin_src = 'https://{{domains[www]}}:{{ports[https][0]}}' +
  same_origin_src;
const header = 'Permissions-Policy header "battery=()"';

promise_test(async t => {
  await promise_rejects_dom(t, 'NotAllowedError', navigator.getBattery());
}, `${header} disallows the top-level document.`);

async_test(t => {
  test_feature_availability('navigator.getBattery()', t, same_origin_src,
      expect_feature_unavailable_default);
}, `${header} disallows same-origin iframes.`);

async_test(t => {
  test_feature_availability('navigator.getBattery()', t, cross_origin_src,
      expect_feature_unavailable_default);
}, `${header} disallows cross-origin iframes.`);
</script>
</body>
